requirements.txt
google-cloud-storage
gradio